Split out some of the logic into selectors/containers
