+++ browser/src/UI/Colors.ts
@@ -1,0 +1,17 @@
+/**
+ * Helper utilities for manipulating colors
+ */
+
+import * as Color from "color"
+
+/**
+ * Gets a reasonable border color for popup elements, based on popups
+ */
+export const getBorderColor = (bgColor: string, fgColor: string): string => {
+    const backgroundColor = Color(bgColor)
+    const foregroundColor = Color(fgColor)
+
+    const borderColor = backgroundColor.luminosity() > 0.5 ? foregroundColor.lighten(0.6) : foregroundColor.darken(0.6)
+    return borderColor.rgb().toString()
+}
+
